+++ swing/framework-ui/src/test/java/is/codion/swing/framework/ui/component/EntityComboBoxTest.java
@@ -76,15 +76,15 @@ public class EntityComboBoxTest {
 		comboBoxModel.select(jonesKey);
 		EntityComboBox comboBox = EntityComboBox.builder(comboBoxModel).build();
 		NumberField<Integer> empIdValue = comboBox.integerSelectorField(Employee.ID).build();
-		assertEquals(3, empIdValue.number().get());
+		assertEquals(3, empIdValue.get());
 		Entity.Key blakeKey = comboBoxModel.connectionProvider().entities().primaryKey(Employee.TYPE, 5);
 		comboBoxModel.select(blakeKey);
-		assertEquals(5, empIdValue.number().get());
+		assertEquals(5, empIdValue.get());
 		comboBoxModel.setSelectedItem(null);
-		assertNull(empIdValue.number().get());
-		empIdValue.number().set(10);
+		assertNull(empIdValue.get());
+		empIdValue.set(10);
 		assertEquals("ADAMS", comboBoxModel.selection().value().get(Employee.NAME));
-		empIdValue.number().set(null);
+		empIdValue.set(null);
 		assertNull(comboBoxModel.selection().value());
 	}
 }